+++ packages/base/src/features/story/components/ListStoryStageOverlay.tsx
@@ -21,48 +21,8 @@ import {
   buildStorySegmentViewItems,
   getStoryMarkdownFromSlide,
 } from '@/src/features/story/utils/storySegmentViewItems';
-import { getSegmentDisplayMode } from '../utils/listStoryScrollTrack';
-
-/**
- * Invokes callback after all images under root have loaded or errored.
- * Returns a cancel function (fix 3 will extract this as a shared helper).
- */
-function whenImagesSettled(
-  root: HTMLElement,
-  callback: () => void,
-): () => void {
-  const images = Array.from(root.querySelectorAll('img'));
-  if (images.length === 0) {
-    callback();
-    return () => {};
-  }
-
-  let cancelled = false;
-  let pending = images.length;
-
-  const settle = (): void => {
-    if (cancelled) {
-      return;
-    }
-    pending -= 1;
-    if (pending <= 0) {
-      callback();
-    }
-  };
-
-  for (const img of images) {
-    if (img.complete) {
-      settle();
-    } else {
-      img.addEventListener('load', settle, { once: true });
-      img.addEventListener('error', settle, { once: true });
-    }
-  }
-
-  return () => {
-    cancelled = true;
-  };
-}
+import { getSegmentDisplayMode } from '@/src/features/story/utils/listStoryScrollTrack';
+import { whenImagesSettled } from '@/src/features/story/utils/whenImagesSettled';
 
 interface IListStoryStageOverlayProps {
   model: IJupyterGISModel;
@@ -230,6 +190,7 @@ export function ListStoryStageOverlay({
 
   useLayoutEffect(() => {
     const parent = overlayRef.current?.parentElement;
+    console.log('SANITY');
 
     if (!parent) {
       setStageHeight(0);
